requirements.txt
flask
opencv-python
Pillow
reportlab
numpy
gunicorn